Move implementations of irol/lrol into BinaryEvaluator.cpp
Previously, the implementations of the irol and lrol opcodes in the
Power codegen were in functions in OMRTreeEvaluator.cpp. This was a
holdover from when these functions were used for inlining certain
recognized methods in OpenJ9. Since this dependency has since been
removed, using separate functions for this is now redundant and the
implementations of the irol and lrol opcodes have been moved into their
respective evaluators.

Signed-off-by: Ben Thomas <ben@benthomas.ca>