Cmake: Fix install of omr_ebcdic and omr_ascii targets
The `install()` command needs to be called from the same directory that
defines the targets. The current code works because the top level
CMakeLists includes  zos.cmake (which defines the libraries). However
If a consumer of omr includes zos.cmake before calling add_subdirectory
on omr, the definition and installation commands will be in different
directories

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>