Fix OMRMonitor.cpp failing to compile on AIX
Previously, OMRMonitor.hpp contained a workaround for an XLC issue with
private delete operators in it. However, this workaround was only used
when compiling on z/OS, meaning that attempting to build the OMR
compiler on AIX with XLC would fail with an error in OMRMonitor.cpp.
This workaround has now been enabled on AIX.

Signed-off-by: Ben Thomas <ben@benthomas.ca>