Make Tril::SimpleCompiler return a function descriptor on AIX
On most platforms, function pointers are simply pointers to the first
instruction in the function. However, on AIX and z/OS this is not true:
they are instead pointers to data structures that themselves contain
such a pointer. Support for calling Tril-compiled functions has now been
added on AIX by adding the ability for Tril::SimpleCompiler to create
such a function descriptor on AIX.

Signed-off-by: Ben Thomas <ben@benthomas.ca>