Skip Tril tests affected by #3525 on Power BE
A number of Tril tests are currently affected by #3525, in which the
code generators spill sub-integer arguments to stack slots incorrectly.
While this works correctly on little endian machines, it does not work
on big endian machines. Previously, such tests were already being
skipped on Z, but this issue also affects the Power codegen when run on
a big-endian machine. In order to allow Tril tests to be run on AIX,
such tests are now skipped when run on Power BE.

Signed-off-by: Ben Thomas <ben@benthomas.ca>