Disable test compiler tests on AIX
The test compiler tests make the assumption that function pointers can
simply be called using the regular call opcodes in the IL. This is not
true on AIX, since function pointers do not actually point to the
address of the first instruction of the function on AIX. As a result,
the test compiler tests will always segfault when run on AIX. However,
since the test compiler tests are effectively deprecated and were never
previously run on AIX in the first place, this is not considered to be
something worth fixing. Thus, the test compiler tests will no longer be
run on AIX when using ctest.

Signed-off-by: Ben Thomas <ben@benthomas.ca>