Remove PPCAlignedLabelInstruction
Previously, the Power codegen had a class of instruction called
PPCAlignedLabelInstruction. This class was meant to be used for labels
that should be aligned on a given boundary, but it was unused. On top of
that, there's no reason to conflate labels and alignment at all: we
could theoretically want to align instructions without adding a label.

For these reasons, this old class of instruction is being removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>