Add isRelocatable boolean inside loadAddressConstant
This boolean can be set by the caller to indicate that the
instructions generated by loadAddressConstant need to be
relocated.

The reason we use this boolean is because the
cg->comp()->compileRelocatableCode() query may not always return the
correct answer for when relocatable code needs to be generated.
For example: a regular JIT compile done remotely. So for now, we
allow the caller to make the decision on if the instructions needs to
be relocatable when not in an AOT compile.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>