Don't do local relocation for remote compilations
Doing a local relocation is problematic because when we do a
relocation on the client, if there is already data in the location
that needs to be patched, then we end up with garbage.
This is becase we expect the patch location to be populated with
zeroes and at patch time we OR the new data in (on Power).

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>