Enlarge SwitchAnalyzer successor counter
The switch analyzer currently uses an int8_t counter to count the number
of time a given block is a successor of the swtich/table. This can easily
overflow and result in undefined behaviour by the C++ code. This change
enlarges the compiler to int32_t to prevent this kind of overflow and
improves the assert to ensure a sesnible number is being generated.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>