Use CG query to check for merge guard support
This commit refactors an x86 codegen query to the common
codegen such that it can be used by the Z codegen as well
inside the supportsMergingGuards() routine. Power and ARM
override this routine to disable this feature as an opt-in.

The allowGuardMerging() routine is also removed as it's
no longer used.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>

Move supportsMergingGuards to the common codegen

This commit also removes allowGuardMerging from OMR

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>

Disable guard merging in Power and ARM

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>