Fix checkAndReplaceRotation 32/64 bit ambiguity
checkAndReplaceRotation tests for pattern validity by comparing an expected
shift amount based on the size of the operand with the constant from the
multiply node:

if ( (1 << correctLeftShiftAmount) != mulConst )
      return false;

Unfortunately, the implicit int constant '1' allows the C++ compiler to ignore
the possibility of shift amounts>32 (undefined behaviour), resulting in 1<<n
giving the same answer as 1<<(n+32). This allowed the simplification to proceed
on a 64 bit operand with shift values that were only valid for 32 bit operands.

The shift has been updated with code crafted by Leonardo which allows the test
to correctly differentiate between 32 and 64 bit valid patterns.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>