Forward format args to trace function in nonfatal assertion
TR::Assertion is passing the format string, but not the variadic
arguments, to fatal assert. This is causing crashes in the printing
helper.

This commit introduces a new function, va_fatal_assertion, which accepts
a va_list argument. Both assertion and fatal_assertion call this helper.

Using this helper, the arguments should be forwarded along correctly to
the printing helper.

Fixes: #4423
Signed-off-by: Robert Young <rwy0717@gmail.com>