Fix mmap with MAP_ANONYMOUS not working on AIX
It turns out that AIX's implementation of the mmap function requires
that the file descriptor be set to -1 when using MAP_ANONYMOUS.
Previously, there were a couple places that were trying to use
MAP_ANONYMOUS without setting the file descriptor to -1, resulting in
them erroring out. This has been corrected and they should now work
correctly.

Signed-off-by: Ben Thomas <ben@benthomas.ca>