gtest: Do not reference temporaries past their lifetimes
This is an error in gtest, not in OMR.

This same error is repeated in a few places, causing different types of
bad behaviour on AIX, including breaking almost all logging and option
processing.

Rather than bind a const mytype& to a temporary result, copy the result
into a local, which ensures it is safe to access the value later in the
program.

When a reference is bound to a temporary value, that reference is
invalid when the value dies, at the end of the "full expression" that
created it. Using this reference to a temporary in a later expression
results in a kind of "use after destruction" error.

Fixes: #4537
Signed-off-by: Robert Young <rwy0717@gmail.com>