Fix missing conversion in splitPostGRA
There was an issue with splitPostGRA where it was creating temp slot
for the node which was either of Int8 or Int16 type. To maintain
alignment, it needed conversion to Int32 type before storing it into
memory. splitPostGRA was missing this conversion and which was causing
issues on Z where it was storing a byte into memory without conversion
and then loads a byte from that slot causing us to load most significant
byte which contains 0 instead of the original value.
This commit adds the missing conversion piece of code for such nodes.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>