Stop IAP setting site specific information on parm symbol
Site specific information that is propagated from caller can not be
stored into parm symbols as the method can be called at different call
sites with different argument info. One exception is the fixed type info
on the outer most method's arguments, because the arguments exist before
entering the method.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>