Fixes bug found when assigning a new real register
When we assign a new real register to a virtual register (for GPRs), we were not
updating the virtual register to hold its new real register. This led to a
discrepancy between the real->virtual and virtual->real mapping between registers. This
commit ensures that the virtual register now knows of its new real register.

Signed-off-by: Pushkar Bettadpur <pushkar.bettadpur@gmail.com>