Don't default initialize isRelocatable boolean
This boolean will now be used to represent if a relocation
is needed for remote as well as AOT compiles. Doing so consolidates
the two questions that the routine needs to answer. Since OMR is
not aware of remote compilations, the decision on whether a relocation
is required will not have to be made explicitly by the caller.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>