Share implementation to re-reserve trampolines on code cache switch
Each code generator duplicates logic to re-reserve a trampoline for a
call target if the code cache is switched between the time the trampoline
is first reserved (likely instruction selection) and binary encoding.

The logic is nearly the same in all cases.  Provide a common version
that is shared by all code generators.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>