Fix JNI performance regression
This is to check in the OMR portion ...

It was caused by changes committed for OpenJ9 issue #4893. When JNILinkage is not
set on the method symbol anymore, OutOfLine performCall eventully evaluated it
through PrivateLinkage and stuck in interpreter. The fundamental reason is that
a new node is created for performCall but the isPreparedForJNI flag is not copied
over and performCall doesn't get the linkage right.

Resolve: #4132

Signed-off-by: Julian Wang <zlwang@ca.ibm.com>