AArch64: Fix code generation for adr and adrp instructions
This commit fixes the wrong encoding of the immediate value for "adr"
and "adrp" instructions.
Those two instructions encode the lower 2 bits of the immediate value
in bits 29-30 of the instruction word, and the upper 19 bits of the
immediate value in bits 5-23 of the instruction word.
There was a mistake with masking and shifting the upper 19 bits.

Signed-off-by: KONNO Kazuhiro <konno@jp.ibm.com>