Add new overload for loadAddressConstant
This new overload allows the consumer of the API to determine
if a relocation record is needed. The overload is added because
it will allow all consumers of this API
to use the new definition of loadAddressConstant, while
not breaking anything in downstream projects. Once the
new definition has been accepted by downstream projects,
they can begin to use the new API as well since if they choose
to do so.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>