Avoid buffer overflow in count_threads()
strncpy() does not add a NUL-terminator if not part of the specified
segment of the source, so if the file name was more than 11 characters
long, buf might not be NUL-terminated and the following strcat()
may write beyond the end of the buffer.

This was highlighted by a warning issued by gcc 9.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>