Ensure strings are NUL-terminated in omrosdump_helpers
strncpy() does not add a NUL-terminator if not part of the specified
segment of the source, so the result may not be NUL-terminated; use
OMR's str_printf() instead. Affected functions are deriveCoreFileName()
and renameDump().

This was highlighted by a warning issued by gcc 9.

Also fixed typo in corePatternFormat parameter name.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>