Add query in Linkage to get target from comp
In order to allow the target to be acquired in
x/codegen/OMRLinkage.hpp, this commit adds a static query to get the
target from the compilation object (which has only been forward declared
in this header). It may not be a good idea to include the Compilation
class headers in case of future circular dependancies.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>