Move the calls to AC_PROG_CC/CXX higher
Autoconf has a rudimentary macro-prerequisite feature, which allows
macros and tests to automatically invoke dependencies.

Because of this system, AC_PROG_CC is being automatically invoked in the
middle of a check for the size of void*, which is much earlier than
where we manually invoke the macros. For clarity, I'm moving the calls
to AC_PROG_CC/CXX higher, to the beginning of the "build environment"
section of the script, and a few lines before the sizeof_void_p test.

Signed-off-by: Robert Young <rwy0717@gmail.com>