Don't use needRelocationsForStatics query when redundant
In some situations, this query was being used in addition
to another query that would also return the same answer. Since
it doesn't make sense to use this query anymore (as the other
one is more appropriate given the data type), this commit
removes uses of such queries.

Additionally, use of the new needRelocationsForCurrentMethodPC
query is added for symbol->isCompiledMethod data types.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>