Redefine CG queries to return correct answer
The queries modified in this PR are used to replace the
generic compileRelocatableCode() query so that it is easier
for downstream projects to override them on a case-by-case
basis. Since the queries are replacing the original
compileRelocatableCode() call, they should be returning the
same result as before in OMR. This commit fixes that issue.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>