Fixes incorrect register copy and GPR0 use in memory reference instructions
When `assignBestRegisterSingle` method looks to perform a register shuffle
on memory reference registers, it checks if the instruction defines the
register in question and if not, places the register copy above the instruction
in question. This is functionally incorrect as registers freed as part of a shuffle
can be allocated elsewhere and subsequently get clobbered as part of the
register copy.

This commit removes the erroneous modification to register copy location and
further, prevents the use of GPR0 across all registers in an instruction
involving a memory reference.

Signed-off-by: Pushkar Bettadpur <pushkar.bettadpur@gmail.com>