Ensured calls to findBestFreeRegister provide a Register Mask and modify check for useGPR0
The `findBestFreeRegister` API provides an argument to set what registers
are available/unavailable for allocation. However a lot of invocations do not
specify this argument and instead, a check is performed within `findBestFreeRegister`
for a memory reference register and found, GPR0 is made unavailable. These checks
lack the appropriate context and it is conceptually cleaner for the caller to set
what registers it would like used.

This commit finds invocations of `findBestFreeRegister` and sets the available
registers field. It also removes the now unnecessary memory reference check.

Signed-off-by: Pushkar Bettadpur <pushkar.bettadpur@gmail.com>