Fixes https://github.com/eclipse/openj9/issues/8020
- MarkMap init consolidated

MarkMap init logic has been merged to resolve the timing hole outlined
in https://github.com/eclipse/openj9/issues/8020. MarkMap init is
dependent on Concurrent State of the global collector, with the init
logic split between `heapAddRange` and `heapReconfigured`, any changes
in states between the two may leave expanded mark map uninited. Hence,
the decision to init (`setMarkBitsInRange`) the mark map has been moved
from `heapAddRange` to `heapReconfigured` which is where we determine if
an update to the init table is required
(`tuneToHeap`/`determineInitWork`). As a result, init is no longer
affected by the change in Concurrent state and we eliminate the timing
window. This guarantee that the MarkMap will be inited, either on the
spot or afterwards by updating the init table.

- Introduced HeapReconfigReason and reworked heapReconfigred API

heapReconfigred now distinguishes different reasons for reconfiguration
(Expand, contract, etc), this functionality is required for the Mark Map
init changes, specifically for Gencon and it is used when Scavenger
resizes tenure _(PhysicalSubArenaVirtualMemorySemiSpace)_ or global
collecter preforms a resize _(PhysicalSubArenaVirtualMemoryFlat)_. At
the time, this is new information is only consumed by Concurrent Global
Collecter (ConcurrentGC.cpp), for policies making use of other
collectors the HeapReconfigReason param defaults to `NONE`.

The Reconfig reasons are dealt in the following ways by Concurrent
Global Collector:

- We should never end up in `heapReconfigured` with `RECONFIG_NONE`
reason
- `RECONFIG_CONTRACT` signifies that `heapRemoveRange` had taken place,
in which case, we just need to update init table (`tuneToHeap`) when
Concurrent is Off, otherwise just `adjustTraceTargets`
- If `heapAddRange` takes place then `heapReconfigured` is called with
`RECONFIG_EXPAND` and have have two different cases:
      1)  heapAddRange was successful (return true), heap reconfig
should be provided with `lowAddress` & `highAddress` which will be used
to init mark map
      2) heapAddRange returns false, signifies a failed `heapAddRange`,
address params are expected to be NULL, in which case mark map won't be
inited but we'll still either `tuneToHeap` or `adjustTraceTarget`.

Signed-off-by: Salman Rana <salman.rana@ibm.com>