Add unit tests for the Power binary encoder
Previously, there was no direct testing of the Power codegen's binary
encoder. While portions of it would be exercised by the Tril tests and
through the test suites of downstream projects such as OpenJ9, there was
a significant amount of functionality that was not tested since it is
either currently unused or only used in very exotic situations that are
unlikely to be exercised by the existing test suite.

In order to better confirm that the Power binary encoder is working as
expected, a new set of unit tests are being added that should exercise
as much of its functionality as possible. The iteration of these tests
being contributed include comprehensive tests for most non-memory
opcodes currently in use by the Power codegen in either OMR or OpenJ9.

Note that a number of these tests are currently disabled due to known
issues in the Power binary encoder. These will be fixed as part of an
ongoing refactor to make the Power binary encoder easier to maintain.

Signed-off-by: Ben Thomas <ben@benthomas.ca>