Add a new dynamic constant API to dereference the pointer
With `JITServer`, the dynamic constant object uintptrj_t
needs to be dereferenced at the client instead of the server.
It’s taken care of by `TR_ResolvedJ9JITServerMethod::dynamicConstant()`.
To prevent the caller from dereferencing the pointer directly,
at the server, adding a new dynamicConstant API to deference it.
The server will send one message to retrieve both the pointer
and the object values in `TR_ResolvedJ9JITServerMethod::dynamicConstant()`.
Since the caller could be run at the client or the server,
the new API should always be used when dereferencing the pointer.
The old API `dynamicConstant(int32_t cpIndex)` will be deprecated next.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>