Fix linkage register preservation in prologue on z/OS
The z/OS system linkage uses the `saveArguments` function to offload
arguments to the stack if needed. This function traverses the parameter
list looking for arguments which are part of the linkage (linkage
arguments) and if needed it stores them to the stack for use in the
method.

This is currently not done on z/OS for Doubles. What ends up happening
is that the linkage register index (lri) is -1 for Double parameters
and as such we never store them off to the stack. Subsequently the
method body when referencing such parameters tries to load them from
the stack and we get back garbage.

To fix this we simply enable the logic to mark Double linkage argument
registers with proper indices on z/OS.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>