Change how far conditional branch expansion works on Power
On Power, the maximum distance of a conditional branch is less than the
maximum distance of an unconditional branch. As a result of this, it is
possible for the target of a conditional branch instruction to be
out of range. In this case, it is necessary to transform that
conditional branch into a flipped conditional branch around an
unconditional branch.

The previous implementation of this had several notable issues. The
expansion of conditional branches was previously delayed until
generateBinaryEncoding was called on the conditional branch in question,
resulting in that function being much more complex than it needs to be.
Additionally, the previous implementation was trying to be clever and
avoid including the extra 4 bytes of estimated length for conditional
branches in some cases, meaning that the estimated locations of
instructions/labels would actually be incorrect and any code using them
needed to manually account for potential expansion of conditional
branches.

This has now been replaced with a new implementation, which performs the
actual expansion prior to calling generateBinaryEncoding by expanding
into multiple TR::Instruction objects. Additionally, the cleverness
relating to estimated lengths of conditional branch instructions has
been removed to simplify how far conditional branches are identified.

Signed-off-by: Ben Thomas <ben@benthomas.ca>