Fix null pointer deref when encoding a MemoryReference
Previously, there was an issue when encoding a TR::MemoryReference on
Power that used delayed indexed form but without an index register set.
The comment in the code implied that the correct thing to do here was to
use addi to add the displacement to the base register. However, instead
we were trying to fill one of the fields of the instruction with the
index register, which was NULL. This code has now been corrected to
encode the correct instruction.

Signed-off-by: Ben Thomas <ben@benthomas.ca>