Fix JitBuilder and TestCompiler JIT shutdown sequence
The shutdown sequence for the compilation controller was never called.
On platforms such as z/OS this is problematic because JitBuilder and
Tril initialize and shutdown the JIT many times, each time leaking
resources in the compilation controller.

One such resource is the thread local TLS `OMR::compilation` variable
which is typically accessed through `TR::comp()`. This TLS variable was
not being properly freed with every shutdown, eventually resulting in
inability to allocate the TLS variable which causes crashes later on.

The JIT compiler currently assumes `TR::comp()` is always non-NULL in
places where it should be available, thus we add fatal asserts if we
are for some reason unable to allocate this variable. These asserts
would have made investigating this problem trivial.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>