Track the total usage of the code cache
To support `-XX:+PrintCodeCache`(eclipse/openj9#8184), the high water
mark of the code cache usage needs to be tracked.
The usage is tracked when increasing/decreasing the code cache repository
take place and also the reuse of the free blocks in the code cache.

The code cache usage reported in the javacore dump file will
reflect this change because increase/decreaseFreeSpaceInCodeCacheRepository
will also be called when the free blocks are reused/freed.

Added `TR_PrintCodeCacheUsage` to `TR_CompilationOptions`. It's set
when `-XX:+PrintCodeCache` is passed to the JVM or when it's set
as an environment variable.

Also added information on the config size, the total free size,
and the used size for each code cache in `printOccupancyStats()`
which is used by OpenJ9 to show the code cache usage
when `TR_PrintCompMem` is set.

Related to eclipse/openj9#8184

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>