Fix generateS390CompareAndBranchOpsHelper
There was a bug in helper function to generate optimized
compare and branch instruction on IBM Z codegen.
When one of the child of `ifacmp*` node is VFT Load from
object which is treated specially on 64-Bit platform with compressed
references and other child is loading value from a temp slot from Stack,
evaluator was generating an instruction to compare only 4 bytes.
This is incorrect on 64-Bit platform as the address load from temp slot on the
stack takes up 8 bytes in the memory.
This commit fixes this issue by identifying such cases and setting up correct
data type for the operation.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>