Fix stack argument passing on Linux on Z and z/OS
Because the Z architecture is big-endian we need to spill the entire
GPR contents for incoming arguments in the register to the stack, since
our stack slots on 64-bit are 64-bit wide. The parameters used within
the method are then properly initialized with offsets into the stack
slot. For example for a 4-byte parameter, we will spill all 8 bytes on
entry if needed and initialize the offset of the parameter with a +4
increment, so that the trees which end up using the parameter within
the method will only access the clean bits of the argument.

This also works for sub-32-bit types as well which more or less fixes
the issues noted in #3525.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>