Rudimentary support for RIP-relative symbols
Provide the means to specify the use of RIP-relative data addressing for
directly accessible symbols whose addresses are known at compile-time.
The current means is to load the target address into a register and then
indirect load/store from the register.

This commit repurposes an existing boolean flag `_preferRIPRelative` that was
not used in OMR nor any downstream project.  It provided some means of suggesting
that RIP-relative addressing is preferred, but it still wasn't guaranteed to
be used.  This boolean was repurposed to now require that RIP-relative addressing
be used if requested.

This implementation provides two new constructors for MemoryReferences based
on those that take symbol references as parameters.  The constructors now
allow the consumer to specify (require) that RIP-relative addressing be used
to access the symbol.

Note that there was already limited support for RIP-relative addressing for
constant data snippets and labels.  That support seems to have been added as
a special case.  It remains in place with this commit, and a much larger
refactoring effort would be required to merge them.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>