Remove duplicate if branches from addConstantToLong
Previously, the addConstantToLong function used on 32-bit performed two
separate checks to see if an addic instruction could be used to perform
the lower part of the addition with duplicate code. The duplicate
branches of this if statement have now been merged together with one
check for the range of lowValue.

This has been done because the previous code was optimized incorrectly
by XLC 13.1 on ppc32 Linux under certain conditions and the range check
would not work correctly, resulting in an invalid addic instruction
being generated.

Signed-off-by: Ben Thomas <ben@benthomas.ca>