Allow extra context to be printed on assertion failures
Previously, assertion failures would only print the message provided to
them and the stack trace at which the assertion occurred. However, it is
often useful to have much more detailed contextual information to allow
simple failures to be debugged without needing to rely on analysis of
corefiles. However, adding all this information to the assert message
would cause these messages to become long and unwieldy and would likely
result in inconsistent formatting of this information.

To resolve this in the most common cases, two new assertion macros have
been added. TR_ASSERT_FATAL_WITH_NODE will print the treetop containing
the provided node on assertion failure. TR_ASSERT_FATAL_WITH_INSTRUCTION
will print the provided instruction and some surrounding instructions,
as well as printing the treetop containing the node that generated the
instruction if such information is available. For both of these macros,
setting the environment variable TR_AssertFullContext will cause the
entire method's trees/instructions to be printed on stderr, which is
useful in scenarios where an assertion failure is not reproducible with
tracing enabled and where jitdumps are not available.

No asserts have yet been modified to use these new macros, but future
code should use them where appropriate to assist in debugging.

Signed-off-by: Ben Thomas <ben@benthomas.ca>