Remove old aliases for record-form VMX instructions
As part of a previous pull request, the VMX record-form instructions
were renamed to be more consistent with other record-form instructions.
However, aliases had to be left for the old names since OpenJ9 still had
code which used the old names. Recently these references were finally
removed, so it should now be safe to remove the old aliases.

Signed-off-by: Ben Thomas <ben@benthomas.ca>