Fix a number of out-of-range immediates on Power
There are a number of places in the Power codegen where we were relying
on the previous behaviour of the binary encoder to ignore the upper bits
of the immediate in an instruction with a 16-bit immediate. The new
binary encoder will now be enforcing that all immediates should be
within the range of values actually representable in the instruction and
will now trigger an assertion failure if an immediate is out of range.

As a result of this, all known code that was relying on the previous
behaviour has been updated to ensure that the immediates they place on
their instructions are within range.

Signed-off-by: Ben Thomas <ben@benthomas.ca>