Remove old formats from opcode properties
Previously, the Power instruction opcode properties table included
formats that were based on the Power ISA's names for the instruction
forms. However, these names are inconsistent and are not generally
useful for knowing the actual format of an instruction. For instance,
the Power ISA details over 90 different encodings for X-form
instructions and the one being used depends on the instruction.

In an effort to make the binary encoder cleaner, easier to understand,
and less buggy, these formats will soon be replaced by formats that
control how the binary encoder fills in fields on instructions. In order
to allow this, the previous formats are being removed for the time
being.

Signed-off-by: Ben Thomas <ben@benthomas.ca>