Remove dynamicConstant(int32_t cpIndex)
PR #4726 adds a new `dynamicConstant()` API
`dynamicConstant(int32_t cpIndex, uintptrj_t *obj)`.
The new API dereferences the dynamic constant object pointer
into the passed in parameter `obj`. The reason is that
in the JITServer mode, the server will need to send a message
to the client to retrieve both the pointer and the object value
in TR_ResolvedJ9JITServerMethod::dynamicConstant().
Since the caller of dynamicConstant () could be run at the
client or the server, the new API should always be used
when the dynamic constant object pointer needs to be dereferenced.

It'd be a better practice to deprecate
dynamicConstant(int32_t cpIndex) to reduce the chances
of the caller from dereferencing the pointer directly at the server.

Closes #4731

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>