Refactor TR::Instruction::generateBinaryEncoding
As the first step towards restructuring the Power binary encoder
implementation on the other instruction types, a baseline implementation
on TR::Instruction must be cleaned up. The generateBinaryEncoding method
on this class has now been split in two. The generateBinaryEncoding
method contains all the boilerplate that will be common to all
instructions, such as copying the template instruction into the
instruction stream and updating the necessary fields. The actual work of
filling in fields will now be performed by a separate
fillBinaryEncodingFields method, which will be overridden by each
instruction class.

Signed-off-by: Ben Thomas <ben@benthomas.ca>