Add asserts to label-relative relocations
Previously, the functions for applying label-relative relocations did
not perform sanity checks such as verifying that the label is in range.
This could result in bugs in binary length estimation and other places
in the binary encoder exhibiting erratic and difficult to debug
behaviour. New asserts have been added to ensure that bugs relating to
label-relative relocations should result in more consistent assertion
failures.

Signed-off-by: Ben Thomas <ben@benthomas.ca>